fix: move timeseries plot to a external method
