+++ src/pandas_profiling/profile_report.py
@@ -32,7 +32,6 @@ from pandas_profiling.report.presentation.flavours.html.templates import (
 from pandas_profiling.serialize_report import SerializeReport
 from pandas_profiling.utils.dataframe import hash_dataframe
 from pandas_profiling.utils.paths import get_config
-from pandas_profiling.visualisation.plot import plot_timeseries_heatmap
 
 
 class ProfileReport(SerializeReport, ExpectationsReport):
@@ -138,77 +137,6 @@ class ProfileReport(SerializeReport, ExpectationsReport):
             )
         else:
             return df
-
-    def _timeseries_heatmap(
-        self,
-        dataframe: pd.DataFrame,
-        entity_column: str,
-        sortby: Optional[Union[str, list]] = None,
-        max_entities: int = 5,
-        selected_entities: Optional[List[str]] = None,
-    ):
-        if sortby is None:
-            sortbykey = "_index"
-            df = dataframe[entity_column].copy().reset_index()
-            df.columns = [sortbykey, entity_column]
-
-        else:
-            if isinstance(sortby, str):
-                sortby = [sortby]
-            cols = [entity_column, *sortby]
-            df = dataframe[cols].copy()
-            sortbykey = sortby[0]
-
-        if df[sortbykey].dtype == "O":
-            try:
-                df[sortbykey] = pd.to_datetime(df[sortbykey])
-            except Exception as ex:
-                raise ValueError(
-                    f"column {sortbykey} dtype {df[sortbykey].dtype} is not supported."
-                ) from ex
-        nbins = np.min([50, df[sortbykey].nunique()])
-
-        df["__bins"] = pd.cut(
-            df[sortbykey], bins=nbins, include_lowest=True, labels=range(nbins)
-        )
-
-        df = df.groupby([entity_column, "__bins"])[sortbykey].count()
-        df = df.reset_index().pivot(entity_column, "__bins", sortbykey).T
-        if selected_entities:
-            df = df[selected_entities].T
-        else:
-            df = df.T[:max_entities]
-        
-        return plot_timeseries_heatmap(self.config, df)
-
-    def timeseries_heatmap(
-        self,
-        dataframe: pd.DataFrame,
-        entity_column: str,
-        sortby: Optional[Union[str, list]] = None,
-        max_entities: int = 5,
-        selected_entities: Optional[List[str]] = None,
-    ):
-        """Generate a multi entity timeseries heatmap based on a pandas DataFrame.
-
-        Args:
-            dataframe: the pandas DataFrame
-            entity_column: name of the entities column
-            sortby: column that define the timesteps (only dates and numerical variables are supported)
-            max_entities: max entities that will be displayed
-            selected_entities: Optional list of entities to be displayed (overules max_entities)
-        """
-    
-        ax = self._timeseries_heatmap(
-            dataframe,
-            entity_column,
-            sortby,
-            max_entities,
-            selected_entities,
-        )
-        # workaround to deal with matpotlib erros with the aspect ratio config
-        ax.set_aspect("equal")
-        return ax
 
     def invalidate_cache(self, subset: Optional[str] = None) -> None:
         """Invalidate report cache. Useful after changing setting.
