+++ src/pandas_profiling/visualisation/plot.py
@@ -3,6 +3,7 @@ import copy
 from typing import Any, Callable, List, Optional, Tuple, Union
 
 import matplotlib
+from matplotlib.axes import Axes
 import numpy as np
 import pandas as pd
 import seaborn as sns
@@ -513,19 +514,56 @@ def plot_acf_pacf(config: Settings, series: pd.Series, figsize: tuple = (15, 5))
     return plot_360_n0sc0pe(config)
 
 
-def plot_timeseries_heatmap(
-    config: Settings, df: pd.DataFrame, figsize: Tuple[int, int] = (12, 5)
+def _prepare_heatmap_data(
+    dataframe: pd.DataFrame,
+    entity_column: str,
+    sortby: Optional[Union[str, list]] = None,
+    max_entities: int = 5,
+    selected_entities: Optional[List[str]] = None,
+) -> pd.DataFrame:
+    if sortby is None:
+        sortbykey = "_index"
+        df = dataframe[entity_column].copy().reset_index()
+        df.columns = [sortbykey, entity_column]
+
+    else:
+        if isinstance(sortby, str):
+            sortby = [sortby]
+        cols = [entity_column, *sortby]
+        df = dataframe[cols].copy()
+        sortbykey = sortby[0]
+
+    if df[sortbykey].dtype == "O":
+        try:
+            df[sortbykey] = pd.to_datetime(df[sortbykey])
+        except Exception as ex:
+            raise ValueError(
+                f"column {sortbykey} dtype {df[sortbykey].dtype} is not supported."
+            ) from ex
+    nbins = np.min([50, df[sortbykey].nunique()])
+
+    df["__bins"] = pd.cut(
+        df[sortbykey], bins=nbins, include_lowest=True, labels=range(nbins)
+    )
+
+    df = df.groupby([entity_column, "__bins"])[sortbykey].count()
+    df = df.reset_index().pivot(entity_column, "__bins", sortbykey).T
+    if selected_entities:
+        df = df[selected_entities].T
+    else:
+        df = df.T[:max_entities]
+    
+    return df
+
+
+def _create_timeseries_heatmap(
+    df: pd.DataFrame,
+    figsize: Tuple[int, int] = (12, 5),
+    color: str = "#337ab7",
 ) -> plt.Axes:
-    """Plot a time series heatmap plot and return the AxesSubplot object.
-    Args:
-        series: The data to plot
-        figsize: The size of the figure (width, height) in inches, default (6,4)
-    Returns:
-        The TimeSeries heatmap.
-    """
     _, ax = plt.subplots(figsize=figsize)
     cmap = matplotlib.colors.LinearSegmentedColormap.from_list(
-        "report", ["white", config.html.style.primary_color], N=64
+        "report", ["white", color], N=64
     )
     pc = ax.pcolormesh(df, edgecolors=ax.get_facecolor(), linewidth=0.25, cmap=cmap)
     pc.set_clim(0, np.nanmax(df))
@@ -534,4 +572,38 @@ def plot_timeseries_heatmap(
     ax.set_xticks([])
     ax.set_xlabel("Time")
     ax.invert_yaxis()
+    return ax
+
+
+def timeseries_heatmap(
+    dataframe: pd.DataFrame,
+    entity_column: str,
+    sortby: Optional[Union[str, list]] = None,
+    max_entities: int = 5,
+    selected_entities: Optional[List[str]] = None,
+    figsize: Tuple[int, int] = (12, 5),
+    color: str = "#337ab7",
+) -> plt.Axes:
+    """Generate a multi entity timeseries heatmap based on a pandas DataFrame.
+
+    Args:
+        dataframe: the pandas DataFrame
+        entity_column: name of the entities column
+        sortby: column that define the timesteps (only dates and numerical variables are supported)
+        max_entities: max entities that will be displayed
+        selected_entities: Optional list of entities to be displayed (overules max_entities)
+        figsize: The size of the figure (width, height) in inches, default (10,5)
+        color: the primary color, default '#337ab7'
+    Returns:
+        The TimeSeries heatmap.
+    """
+    df = _prepare_heatmap_data(
+        dataframe,
+        entity_column,
+        sortby,
+        max_entities,
+        selected_entities
+    )
+    ax = _create_timeseries_heatmap(df, figsize, color)
+    ax.set_aspect(1)
     return ax
